requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
bm25s==0.1.10
boto3==1.34.144
botocore==1.34.144
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.8
Cython==3.0.10
dacite==1.8.1
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
duckduckgo_search==6.1.7
faiss-cpu==1.8.0.post1
fastapi==0.109.2
fastavro==1.9.5
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fschat==0.2.36
fsspec==2024.6.1
ftfy==6.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
latex2mathml==3.77.0
lightgbm==4.4.0
llama-cloud==0.0.6
llama-index==0.10.54.post1
llama-index-agent-openai==0.2.6
llama-index-cli==0.1.12
llama-index-core==0.10.54
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.4
llama-index-legacy==0.9.48
llama-index-llms-groq==0.1.4
llama-index-llms-openai==0.1.25
llama-index-llms-openai-like==0.1.3
llama-index-multi-modal-llms-openai==0.1.7
llama-index-postprocessor-cohere-rerank==0.1.7
llama-index-postprocessor-rankllm-rerank==0.1.3
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-database==0.1.4
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-index-readers-web==0.1.23
llama-index-retrievers-bm25==0.2.0
llama-index-tools-wikipedia==0.1.2
llama-index-vector-stores-pinecone==0.1.7
llama-parse==0.4.7
loguru==0.7.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.18
nltk==3.8.1
nmslib==2.0.6
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
peft==0.11.1
pillow==10.4.0
pinecone-client==3.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
pyarrow==17.0.0
pybind11==2.13.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyjnius==1.6.1
pypdf==4.2.0
pyreqwest_impersonate==0.4.9
pyserini==0.36.0
PyStemmer==2.2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
psycopg2-binary==2.9.9
rank-llm==0.12.8
regex==2024.5.15
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.36.3
striprtf==0.0.26
svgwrite==1.4.3
sympy==1.13.0
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
torch==2.2.2
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.2
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
llama-index-vector-stores-qdrant==0.2.14
qdrant-client==1.10.1
traceloop-sdk==0.26.2
opentelemetry-instrumentation-sqlalchemy==0.47b0
llama-index-readers-json==0.1.5
llama-index-postprocessor-rankgpt-rerank==0.1.6
cohere==5.8.0